Synchronize listeners array
This makes the listeners array list interactions thread-safe. Currently, if a listener is added or removed while receiving a notification you will get a concurrent modification exception. This fixes that by simply wrapping the listeners in a synchronized list impl and synchronizing on the listeners array any time iteration is to be performed.